Fix RDL analysis: add PersonLocker to existing file, remove duplicate
- Apply Fix 4 (PersonLocker) directly to romanian_deadlift_analysis.py
- PersonLocker tracks locked person's hip centroid, rejects detections
  that jump >60% of body size to prevent skeleton snapping to bystanders
- Draw last_valid_lm skeleton on frames where pose is lost or rejected
- Remove romanian_deadlift_analysis_fixed.py (changes belong in the main file)
